refactor(command): 重构命令模块结构并合并matcher和handler功能
- 将puniyu_matcher_command和puniyu_handler_command合并为puniyu_command
- 移除独立的matcher和handler模块，整合到单一command模块中
- 更新Cargo.lock和依赖配置以反映模块合并
- 重命名CooldownScope中的GroupUser为GroupMember
- 重构命令匹配器实现，优化文本提取逻辑
- 添加命令处理的参数解析和错误处理功能
- 实现命令权限检查和冷却期控制机制
