+++ .release-please-manifest.json
@@ -6,8 +6,7 @@
   "crates/puniyu_config": "0.6.0",
   "crates/puniyu_types": "0.6.0",
   "crates/puniyu_registry": "0.6.0",
-  "crates/puniyu_matcher_command": "0.6.0",
-  "crates/puniyu_handler_command": "0.6.0",
+  "crates/puniyu_command": "0.6.0",
   "crates/puniyu_bus": "0.6.0",
   "crates/puniyu_server": "0.4.1",
   "packages/puniyu_plugin": "0.6.0",
