+++ crates/puniyu_bus/Cargo.toml
@@ -12,12 +12,10 @@ rust-version.workspace = true
 
 [dependencies]
 tokio = { workspace = true, features = ["macros"] }
-puniyu_logger.workspace = true
 
 puniyu_types = { version = "0.6.0", path = "../puniyu_types", default-features = false, features = [
   "bus",
   "matcher",
   "handler",
 ] }
-puniyu_matcher_command = { version = "0.6.0", path = "../puniyu_matcher_command" }
-puniyu_handler_command = { version = "0.6.0", path = "../puniyu_handler_command" }
+puniyu_command = { version = "0.6.0", path = "../puniyu_command" }
