--- crates/puniyu_handler_command/CHANGELOG.md
@@ -1,22 +1,0 @@
-# Changelog
-
-## [0.6.0](https://github.com/puniyu/puniyu/compare/handler-command-v0.5.12...handler-command-v0.6.0) (2025-12-02)
-
-
-### ✨ 新功能
-
-* **config:** 添加全局命令前缀配置并优化模块结构 ([e0e5bb5](https://github.com/puniyu/puniyu/commit/e0e5bb5ac9eb24a37189a3d50b42d3db8db58dde))
-
-
-### 🔧 其他更新
-
-* **release:** 切换到 release-plz 发布流程 ([2e2e061](https://github.com/puniyu/puniyu/commit/2e2e06162cc9d953d4bf193669a8ff8babc5e9f5))
-
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * puniyu_matcher_command bumped from 0.1.0 to 0.6.0
-    * puniyu_registry bumped from 0.5.12 to 0.6.0
-    * puniyu_types bumped from 0.5.12 to 0.6.0
