--- crates/puniyu_handler_command/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "puniyu_handler_command"
-version = "0.6.0"
-description = "Puniyu handler command"
-authors.workspace = true
-license.workspace = true
-readme.workspace = true
-homepage.workspace = true
-repository.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-async-trait.workspace = true
-puniyu_logger.workspace = true
-clap = { workspace = true, features = ["string", "error-context"] }
-
-puniyu_matcher_command = { version = "0.6.0", path = "../puniyu_matcher_command" }
-puniyu_registry = { version = "0.6.0", path = "../puniyu_registry", default-features = false, features = ["command"] }
-puniyu_types = { version = "0.6.0", path = "../puniyu_types", default-features = false, features = ["handler"] }
